+++ CHANGELOG.md
@@ -1,4 +1,14 @@
-## 0.1.0 (Unreleased)
+## 0.2.0
+
+FEATURES:
+
+- Direct translation of terraform values.
+
+FIXES:
+
+- Unmarshalling errors with null values ([#12](https://github.com/isometry/terraform-provider-deepmerge/issues/12))
+
+## 0.1.0
 
 FEATURES: